tests from config_test.go separated into separate files
